+++ src/qwertypy/data_analysis/utils.py
@@ -1,0 +1,4 @@
+import numpy as np
+
+def npArrayReshape(some_array):
+    return np.array(some_array).reshape(-1, 1)