+++ CHANGELOG.md
@@ -5,11 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased (0.4.0)]
 
 ### Added
 
--   `qwertypy.tickertape.financials`: Added `getYearsAndValues`.
+-   `qwertypy.tickertape.financials.getYearsAndValues`.
+-   `qwertypy.data_analysis`.
+-   `qwertypy.data_analysis.regression`.
+-   `qwertypy.data_analysis.regression.QpyLinearRegression`.
 
 ### Internal